// libs/triangle/gridster/src/gridster-config.interface.ts
export enum CompactType {
  None                = 'none',
  CompactUp           = 'compactUp',
  CompactLeft         = 'compactLeft',
  CompactUpAndLeft    = 'compactUp&Left',
  CompactLeftAndUp    = 'compactLeft&Up',
  CompactRight        = 'compactRight',
  CompactUpAndRight   = 'compactUp&Right',
  CompactRightAndUp   = 'compactRight&Up',
  CompactDown         = 'compactDown',
  CompactDownAndLeft  = 'compactDown&Left',
  CompactLeftAndDown  = 'compactLeft&Down',
  CompactDownAndRight = 'compactDown&Right',
  CompactRightAndDown = 'compactRight&Down'
}